Moved ueGear tag functionality into its own module
